Split implementation from the index files for each entity
