+++ cloud/wfm/src/executor/index.ts
@@ -1,12 +1,10 @@
-import * as Promise from 'bluebird';
-import { Process } from '../process';
-import { ProcessInstance } from '../process-instance';
-import { Task, TaskEventData, TaskStatus } from '../task';
+import {Process} from '../process';
+import {ProcessInstance} from '../process-instance';
 
 /**
- * Executor engine for a Process
+ * Executor engine for a {@link Process}
  *
- * Triggers instantiation and execution of a ProcessInstance
+ * Triggers instantiation and execution of a {@link ProcessInstance}
  */
 export interface Executor {
   process: Process;
@@ -14,32 +12,6 @@ export interface Executor {
   start(): void;
 }
 
-export interface InstanceRepository {
-  save(instance: ProcessInstance): Promise<ProcessInstance>;
-}
-
-class ExecutorImpl implements Executor {
-  public instance: ProcessInstance;
-  constructor(public process: Process, public instanceRepository: InstanceRepository) {
-    this.instance = this.process.createInstance();
-  }
-  public start() {
-    this.runCurrentTask();
-  }
-  protected onTaskDone(e: TaskEventData<Task>) {
-    if (e.currentStatus === TaskStatus.done) {
-      Promise.resolve(this.instance.next())
-        .then(() => this.saveInstance)
-        .then(() => this.runCurrentTask());
-    }
-  }
-  protected saveInstance() {
-    return this.instanceRepository.save(this.instance);
-  }
-  protected runCurrentTask() {
-    this.instance.getCurrentTask()
-      .then(t => t.on('statusChange', this.onTaskDone).run());
-  }
-}
-
+import ExecutorImpl, {InstanceRepository} from './ExecutorImpl';
+export {InstanceRepository};
 export default ExecutorImpl;
