+++ cloud/wfm/src/process/index.ts
@@ -1,12 +1,11 @@
-import * as _ from 'lodash';
-import Instance, { ProcessInstance } from '../process-instance';
-import TaskImpl, { Task } from '../task';
+import {ProcessInstance} from '../process-instance';
+import {Task} from '../task';
 /**
- * Definition holder for a set of Tasks
- * Intended to be instantiated as a ProcessInstance  in order to be executed
+ * Definition holder for a linear set of {@link Task}s
+ * Intended to be instantiated as a {@link ProcessInstance} in order to be executed
  */
 export interface Process {
-  /** Unique identifier for this Process */
+  /** Unique identifier for this {@link Process} */
   id: string;
   /** Description for UI */
   displayName: string;
@@ -15,14 +14,5 @@ export interface Process {
   createInstance(): ProcessInstance;
 }
 
-export default class ProcessImpl implements Process {
-  public tasks: Task[];
-  public id: string;
-  constructor(public displayName: string) {
-  }
-
-  public createInstance() {
-    const instance = new Instance(this.tasks);
-    return instance;
-  }
-}
+import ProcessImpl from './ProcessImpl';
+export default ProcessImpl;
