+++ cloud/wfm/src/task/BaseTask.ts
@@ -1,14 +1,27 @@
 import { EventEmitter } from 'eventemitter3';
+import {Result} from '../result';
 import { Task, TaskEventData, TaskStatus } from './index';
 
 /**
- * Base implementation for
+ * Base implementation for {@link Task}s
+ * Derived classes are expected to implement their run {@link #run} methods
  */
 class BaseTask extends EventEmitter implements Task {
+  public result: Result;
+  /**
+   * Storage field for the `status` property
+   */
   protected _status: TaskStatus | number = TaskStatus.pending;
   protected options: object;
 
+  /**
+   * Setter for the status property
+   * Emits a statusChange event when set
+   */
   public set status(to: TaskStatus | number) {
+    if (to === this._status) {
+      return;
+    }
     const e: TaskEventData<this> = {
       date: new Date(),
       previousStatus: this.getStatus(),
@@ -18,6 +31,7 @@ class BaseTask extends EventEmitter implements Task {
     this._status = to;
     this.emit('statusChange', e);
   }
+
   public get status() {
     return this._status;
   }
@@ -25,13 +39,14 @@ class BaseTask extends EventEmitter implements Task {
   /**
    * Start the execution.
    * This method is expected to be overridden in specific implementations
-   * `super()` can be called if just resolving to the {@link TaskStatus#done} value
+   *
+   * This implementation simply sets the status to {@link TaskStatus#done}
    */
   public run() {
     this.status = TaskStatus.done;
   }
 
-  public getOptionSchema = () => ({});
+  public getOptionsSchema = () => ({});
   public setOptions(options: object) {
     // TODO: JsonSchema validation with http://epoberezkin.github.io/ajv/
     this.options = options;
